+++ xfree/xc/programs/twm/util.c
@@ -48,7 +48,7 @@ in this Software without prior written authorization from The Open Group.
 /**    TORTIOUS ACTION, ARISING OUT OF OR IN  CONNECTION  WITH  THE  USE    **/
 /**    OR PERFORMANCE OF THIS SOFTWARE.                                     **/
 /*****************************************************************************/
-/* $XFree86: xc/programs/twm/util.c,v 1.10 2001/12/14 20:01:11 dawes Exp $ */
+/* $XFree86: xc/programs/twm/util.c,v 1.13 2002/09/24 21:00:28 tsi Exp $ */
 
 
 /***********************************************************************
@@ -295,8 +295,9 @@ char *name;
     newname = (char *) malloc (HomeLen + strlen(name) + 2);
     if (!newname) {
 	fprintf (stderr, 
-		 "%s:  unable to allocate %d bytes to expand filename %s/%s\n",
-		 ProgramName, HomeLen + strlen(name) + 2, Home, &name[1]);
+		 "%s:  unable to allocate %ld bytes to expand filename %s/%s\n",
+		 ProgramName, HomeLen + (unsigned long)strlen(name) + 2,
+		 Home, &name[1]);
     } else {
 	(void) sprintf (newname, "%s/%s", Home, &name[1]);
     }
@@ -447,8 +448,8 @@ InsertRGBColormap (a, maps, nmaps, replace)
     if (!sc) {				/* no existing, allocate new */
 	sc = (StdCmap *) malloc (sizeof (StdCmap));
 	if (!sc) {
-	    fprintf (stderr, "%s:  unable to allocate %d bytes for StdCmap\n",
-		     ProgramName, sizeof (StdCmap));
+	    fprintf (stderr, "%s:  unable to allocate %ld bytes for StdCmap\n",
+		     ProgramName, (unsigned long)sizeof (StdCmap));
 	    return;
 	}
     }
